`pval` is part of `MonoPhenotypeAnalysisResult`.
